Move usersession to core.
