+++ src/main/java/seedu/clinicio/commons/events/ui/LoginSuccessEvent.java
@@ -1,7 +1,7 @@
 package seedu.clinicio.commons.events.ui;
 
 import seedu.clinicio.commons.events.BaseEvent;
-import seedu.clinicio.commons.util.UserSession;
+import seedu.clinicio.commons.core.UserSession;
 
 /**
  * Indicate the login is successful.
